+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # VRFSearchTool.py #
 ---
 
+## VRFSearchTool v0.0.16-beta (2013-08-29) ##
+* Added basic logging to file to track results if application has to connect
+  to a router to run buildIndex().
+* Suppressed error SPAM from stdout by adding stderr=(open(os.devnull, 'w'))
+  to the Queue() function. (Errors are still written to the log.)
+
 ## VRFSearchTool v0.0.15-beta (2013-08-28) ##
 * Added functionality to specify configFile from the command line.
 * Updated README.md